EOS-19072-ImproveCoverage : Coverage Setup script updated and used instead of shell scripts.Shell script files are deleted.

+++ sspl_test/coverage/coverage_setup.py
@@ -15,20 +15,58 @@
 # about this software or licensing, please email opensource@seagate.com or
 # cortx-questions@seagate.com.
 
-# NOTE : This script is written to replace start_sspl_coverage.sh and
-#        stop_sspl_coverage.sh bash scripts with python script in future.
-#        currently this file is not in use.
-
 import shutil
 import os
 import pwd
 import signal
 import psutil
 import time
+import sys
 from cortx.utils.process import SimpleProcess
+# Add the top level directory to the sys.path to access classes
+topdir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+os.sys.path.insert(0, topdir)
+from framework.base.sspl_constants import DATA_PATH
 
 file_path = "/opt/seagate/cortx/sspl/low-level"
-report_path = "/var/cortx/sspl/coverage/sspl_xml_coverage_report.xml"
+report_path = f"{DATA_PATH}coverage/sspl_xml_coverage_report.xml"
+
+PATCH_1 = """\
+from coverage import Coverage
+co = Coverage(
+        data_file='%scoverage/.sspl_coverage_report',
+        include="/opt/seagate/*",
+        omit=['*/.local/*', '*/usr/*'],
+        # config_file='%scoverage/.coveragerc',
+        )\
+"""%(DATA_PATH, DATA_PATH)
+
+PATCH_2 = """\
+        logger.info("Starting coverage report scope")
+        co.start()\
+"""
+
+PATCH_3 = """\
+def generate_cov_report(signal_number, frame):
+    logger.info('Ending Coverage Scope')
+    co.stop()
+    logger.info('coverage object stopped.')
+    co.save()
+    logger.info('coverage info saved.')
+    cov_per = co.xml_report(ignore_errors=True,
+        outfile='%scoverage/sspl_xml_coverage_report.xml')
+    logger.info(f'XML coverage report generated with coverage of {cov_per} percentage.')
+    ## Enable below code to inable HTML report generation
+ 	# html_cov_per = co.html_report(
+ 	#                     directory='%scoverage/sspl_html_coverage',
+ 	#                     ignore_errors=True,
+ 	#                 )
+ 	# logger.info(f'HTML coverage report geverated with coverage of {html_cov_per} percentage.')\
+"""%(DATA_PATH, DATA_PATH)
+
+PATCH_4 = """\
+    signal.signal(signal.SIGUSR1, generate_cov_report)\
+"""
 
 def coverage_setup():
     """Installs pip3 coverage package. Injects different patches code from
@@ -36,49 +74,33 @@ def coverage_setup():
        for code coverage report and assigns permission to the directory.
     """
     print("Installing coverage.py")
-    _, _, return_code = SimpleProcess('python3 -m pip install coverage').run()
+    _, err, return_code = SimpleProcess('python3 -m pip install coverage').run()
     if return_code:
+        print(err)
         return return_code
 
     print("Creating required files for coverage..")
     patch1_name = "initialize coverage obj for code coverage report generation"
-    patch1_start_ind = 0
-    patch1_n_of_lines = 7
-
     patch2_name = "start the code coverage scope"
-    patch2_start_ind = 7
-    patch2_n_of_lines = 2
-
     patch3_name = "stop coverage, save and generate code coverage report"
-    patch3_start_ind = 9
-    patch3_n_of_lines = 15
-
     patch4_name = "signal handler for SIGUSR1 to generate code coverage report"
-    patch4_start_ind = 24
-    patch4_n_of_lines = 1
-
-    if os.path.exists(f'{file_path}/sspl_ll_d_cov'):
-        os.remove(f'{file_path}/sspl_ll_d_cov')
-
-    with open('./coverage_code', 'r') as cov_code:
-        cov_lines = cov_code.readlines()
 
     with open(f'{file_path}/sspl_ll_d', 'r') as sspl_ll_d:
         sspl_ll_d_lines = sspl_ll_d.readlines()
 
     for i, line in enumerate(sspl_ll_d_lines):
         if patch1_name in line:
-            for j in range(patch1_n_of_lines):
-                sspl_ll_d_lines.insert(i+j+1, cov_lines[patch1_start_ind+j])
+            for j,l in enumerate(PATCH_1.split('\n')):
+                sspl_ll_d_lines.insert(i+j+1, l+'\n')
         elif patch2_name in line:
-            for j in range(patch2_n_of_lines):
-                sspl_ll_d_lines.insert(i+j+1, cov_lines[patch2_start_ind+j])
+            for j,l in enumerate(PATCH_2.split('\n')):
+                sspl_ll_d_lines.insert(i+j+1, l+'\n')
         elif patch3_name in line:
-            for j in range(patch3_n_of_lines):
-                sspl_ll_d_lines.insert(i+j+1, cov_lines[patch3_start_ind+j])
+            for j,l in enumerate(PATCH_3.split('\n')):
+                sspl_ll_d_lines.insert(i+j+1, l+'\n')
         elif patch4_name in line:
-            for j in range(patch4_n_of_lines):
-                sspl_ll_d_lines.insert(i+j+1, cov_lines[patch4_start_ind+j])
+            for j,l in enumerate(PATCH_4.split('\n')):
+                sspl_ll_d_lines.insert(i+j+1, l+'\n')
 
     shutil.move(f'{file_path}/sspl_ll_d',
                 f'{file_path}/sspl_ll_d.bak')
@@ -93,11 +115,10 @@ def coverage_setup():
     os.chmod(f"{file_path}/sspl_ll_d", 0o755)
     os.chown(f'{file_path}/sspl_ll_d', uid, -1)
 
-    os.makedirs('/var/cortx/sspl/coverage/', 0o755, exist_ok=True)
-    os.chown('/var/cortx/sspl/coverage/', uid, -1)
+    os.makedirs(f'{DATA_PATH}coverage/', 0o755, exist_ok=True)
+    os.chown(f'{DATA_PATH}coverage/', uid, -1)
 
     return 0
-
 
 def coverage_reset():
     """Send SIGUSR1 signal to sspl_ll_d to trigger code coverage report generation.
@@ -111,17 +132,41 @@ def coverage_reset():
     os.kill(pid, signal.SIGUSR1)
     time.sleep(5)
 
-    modification_time = \
-        os.path.getmtime(report_path)
+    if os.path.isfile(report_path):
+        modification_time = \
+            os.path.getmtime(report_path)
 
-    if (time.time() - modification_time) < 100:
-        modification_time = time.strftime('%Y-%m-%d %H:%M:%S',
-                                          time.localtime(modification_time))
-        print("%s : The Code Coverage Report is saved at %s" %
-              (modification_time, report_path))
+        if (time.time() - modification_time) < 100:
+            modification_time = time.strftime('%Y-%m-%d %H:%M:%S',
+                                        time.localtime(modification_time))
+            print("%s : The Code Coverage Report is saved at %s" %
+                (modification_time, report_path))
+        else:
+            print("The Code Coverage Report generation failed.")
     else:
-        print("The Code Coverage Report is not generated.")
+        print("%s file does not exists."%report_path)
 
-    os.remove(f'{file_path}/sspl_ll_d')
-    shutil.move(f'{file_path}/sspl_ll_d.bak',
-                f'{file_path}/sspl_ll_d')
+    if os.path.isfile(f'{file_path}/sspl_ll_d.bak'):
+        os.remove(f'{file_path}/sspl_ll_d')
+        shutil.move(f'{file_path}/sspl_ll_d.bak',
+                    f'{file_path}/sspl_ll_d')
+
+def print_help():
+    print('Error: Incorrect arguments to coverage_setup file.\n'
+        'cmd : python3 coverage_setup.py [start/stop]\n'
+        'Args => \n'
+        '    start : Set-up the environment for code coverage.\n'
+        '    stop : Reset the sspl environmet to normal.')
+
+if __name__ == "__main__":
+    if len(sys.argv) != 2:
+        print_help()        
+        sys.exit(100)
+    if sys.argv[-1] == 'start':
+        err_code = coverage_setup()
+        sys.exit(err_code)
+    elif sys.argv[-1] == 'stop':
+        coverage_reset()
+    else:
+        print_help()
+        sys.exit(100)
